No need for the proxy mechanism.
