+++ app.py
@@ -1,13 +1,22 @@
-from os import environ
 from ast import literal_eval
-import requests
-from functools import wraps
-from urllib.parse import urlparse, urljoin
+import boto3
+import botocore
 from datetime import datetime, timedelta
+from functools import wraps
+import imghdr
+from os import environ
+import random
+import requests
+import string
+from urllib.parse import urlparse, urljoin
+from werkzeug.utils import secure_filename
 
-from flask import Flask, request, redirect, session, abort, url_for, render_template
+from flask import Flask, request, redirect, session, abort, url_for, render_template, current_app
+from flask_wtf import FlaskForm
+from flask_wtf.file import FileField, FileRequired
+from wtforms.validators import ValidationError
+
 import error_handling
-from proxy import proxy_request
 
 import logging
 
@@ -45,6 +54,7 @@ def setup_logging():
         app.logger.addHandler(logging.StreamHandler())
         app.logger.setLevel(getattr(logging, app.config['LOG_LEVEL']))
 
+
 def login_required(func):
     @wraps(func)
     def handle_login(*args, **kwargs):
@@ -63,6 +73,7 @@ def login_required(func):
             return redirect('{}?scope=read:org&client_id={}'.format(AUTHORIZE_URL, app.config['GITHUB_CLIENT_ID']))
     return handle_login
 
+
 def is_safe_url(target):
     '''
         Ensure a url is safe to redirect to, from WTForms
@@ -73,33 +84,106 @@ def is_safe_url(target):
     return test_url.scheme in ('http', 'https') and \
            ref_url.netloc == test_url.netloc
 
+
+#
+# Mime typing checking, taken straight from Perma (more rigorous than WTForms)
+#
+
+# Map allowed file extensions to mime types.
+# WARNING: If you change this, also change `accept=""` and the label in
+# uploader.html
+file_extension_lookup = {
+    'jpg': 'image/jpeg',
+    'jpeg': 'image/jpeg',
+    'png': 'image/png',
+    'gif': 'image/gif'
+}
+
+# Map allowed mime types to new file extensions and validation functions.
+# We manually pick the new extension instead of using MimeTypes().guess_extension,
+# because that varies between systems.
+mime_type_lookup = {
+    'image/jpeg': {
+        'new_extension': 'jpg',
+        'valid_file': lambda f: imghdr.what(f) == 'jpeg',
+    },
+    'image/png': {
+        'new_extension': 'png',
+        'valid_file': lambda f: imghdr.what(f) == 'png',
+    },
+    'image/gif': {
+        'new_extension': 'gif',
+        'valid_file': lambda f: imghdr.what(f) == 'gif',
+    }
+}
+
+def get_mime_type(file_name):
+    """ Return mime type (for a valid file extension) or None if file extension is unknown. """
+    file_extension = file_name.rsplit('.', 1)[-1].lower()
+    return file_extension_lookup.get(file_extension)
+
+
+def filename_already_used(filename):
+    """Technique from https://stackoverflow.com/a/33843019"""
+    s3 = boto3.resource('s3',
+            aws_access_key_id=current_app.config['AWS_ACCESS_KEY_ID'],
+            aws_secret_access_key=current_app.config['AWS_SECRET_ACCESS_KEY']
+    )
+    exists = False
+    try:
+        s3.Object(current_app.config['S3_BUCKET'], filename).load()
+    except botocore.exceptions.ClientError as e:
+        if e.response['Error']['Code'] == "404":
+            exists = False
+        else:
+            raise
+    else:
+        exists = True
+    return exists
+
+#
+# WTForms custom validators
+#
+
+def valid_mimetype(form, field):
+    mime_type = get_mime_type(field.data.filename)
+    if not mime_type or not mime_type_lookup[mime_type]['valid_file'](field.data):
+        raise ValidationError("Invalid file format")
+
+class UploadForm(FlaskForm):
+    file = FileField(validators=[FileRequired(), valid_mimetype],
+                     label="valid formats: {}".format(", ".join(file_extension_lookup.keys())))
+
+
 ###
 ### ROUTES
 ###
 
-@app.route('/', defaults={'path': ''}, methods=['GET', 'POST'])
-@app.route('/<path:path>')
+@app.route('/', methods=['GET', 'POST'])
 # @login_required
-def catch_all(path):
-    '''
-        All requests are caught by this route, unless explicitly caught by
-        other more specific patterns.
-        http://flask.pocoo.org/docs/0.12/design/#the-routing-system
-    '''
-    def fallback():
-        return render_template('generic.html', context={'heading': "lil blog media uploader",
-                                                        'message': "a static site helper"})
-
-    try:
-        proxied_response = proxy_request(request, path)
-        if proxied_response:
-            return proxied_response
-        else:
-            app.logger.warning("No response returned by proxied endpoint.")
-            return fallback()
-    except NameError:
-        app.logger.warning("No proxy function available.")
-        return fallback()
+def landing():
+    form = UploadForm()
+    if form.validate_on_submit():
+        # Get a safe filename
+        f = form.file.data
+        filename = secure_filename(f.filename)
+        unique_filename = False
+        while not unique_filename:
+            if filename_already_used(filename):
+                fn, ext = os.path.splitext(filename)
+                filename = '{}-{}{}'.format(fn, ''.join(random.choice(string.ascii_uppercase + string.digits) for _ in range(3)), ext)
+                continue
+            unique_filename = True
+        # Upload to s3
+        s3 = boto3.client(
+            's3',
+            aws_access_key_id=current_app.config['AWS_ACCESS_KEY_ID'],
+            aws_secret_access_key=current_app.config['AWS_SECRET_ACCESS_KEY'],
+        )
+        s3.upload_fileobj(f.stream, 'lil-blog-media', filename, ExtraArgs={'ContentType': get_mime_type(filename)})
+        return render_template('success.html', context={'heading': "Your file is up!" ,
+                                                        'url': "https://{}.s3.amazonaws.com/{}".format(current_app.config['S3_BUCKET'], filename) })
+    return render_template('uploader.html', context={'heading': 'Upload Media', 'limit': current_app.config['MAX_CONTENT_LENGTH']//1024//1024}, form=form)
 
 
 @app.route('/login')
